@@ -1,7 +1,9 @@
 package gh2;
 
 // TODO: uncomment the following import once you're ready to start this portion
-// import deque.Deque;
+import deque.ArrayDeque;
+import deque.Deque;
+import deque.LinkedListDeque;
 // TODO: maybe more imports
 
 //Note: This file will not compile until you complete the Deque implementations
@@ -14,7 +16,7 @@ public class GuitarString {
 
     /* Buffer for storing sound data. */
     // TODO: uncomment the following line once you're ready to start this portion
-    // private Deque<Double> buffer;
+     private Deque<Double> buffer;
 
     /* Create a guitar string of the given frequency.  */
     public GuitarString(double frequency) {
@@ -22,6 +24,11 @@ public class GuitarString {
         //       cast the result of this division operation into an int. For
         //       better accuracy, use the Math.round() function before casting.
         //       Your should initially fill your buffer array with zeros.
+        buffer = new ArrayDeque<>();
+        int size = Math.round(SR / (float)frequency);
+        for(int i = 0; i < size; i++){
+            buffer.addLast(0.0);
+        }
     }
 
 
@@ -35,6 +42,12 @@ public class GuitarString {
         //       other. This does not mean that you need to check that the numbers
         //       are different from each other. It means you should repeatedly call
         //       Math.random() - 0.5 to generate new random numbers for each array index.
+        int s = buffer.size();
+        for(int i = 0; i < s; i++){
+            buffer.removeFirst();
+            double r = Math.random() - 0.5;
+            buffer.addLast(r);
+        }
     }
 
     /* Advance the simulation one time step by performing one iteration of
@@ -44,12 +57,15 @@ public class GuitarString {
         // TODO: Dequeue the front sample and enqueue a new sample that is
         //       the average of the two multiplied by the DECAY factor.
         //       **Do not call StdAudio.play().**
+        double now = buffer.removeFirst();
+        double newdouble = DECAY * (now + buffer.get(0)) * 0.5;
+        buffer.addLast(newdouble);
     }
 
     /* Return the double at the front of the buffer. */
     public double sample() {
         // TODO: Return the correct thing.
-        return 0;
+        return buffer.get(0);
     }
 }
     // TODO: Remove all comments that say TODO when you're done.
